Renamed files & moved constants
